1.jacobi match improved 2.const bigint import 3.test unit add 4.elgamal publickkey trait write 5.remove redundant function for elgamal_utils
